Moving error exception to transport.grpc
