+++ openfl/transport/grpc/director_client.py
@@ -10,7 +10,6 @@ from typing import Type
 
 import grpc
 
-from openfl import ShardNotFoundError
 from openfl.interface.interactive_api.shard_descriptor import ShardDescriptor
 from openfl.pipelines import NoCompressionPipeline
 from openfl.protocols import director_pb2
@@ -18,6 +17,7 @@ from openfl.protocols import director_pb2_grpc
 from openfl.protocols import interceptors
 from openfl.protocols.utils import construct_model_proto
 from openfl.protocols.utils import deconstruct_model_proto
+from openfl.transport.grpc.exceptions import ShardNotFoundError
 from openfl.transport.grpc.director_server import CLIENT_ID_DEFAULT
 
 from .grpc_channel_options import channel_options
