+++ openfl/transport/grpc/exceptions.py
@@ -1,0 +1,2 @@
+class ShardNotFoundError(Exception):
+    """Indicates that director has no information about that shard."""
